+++ components/tutors-ui/index.js
@@ -4,7 +4,9 @@ export { default as Card } from './lib/Molecules/Cards/Card.svelte';
 export { default as NoteCard } from './lib/Molecules/Cards/NoteCard.svelte';
 export { default as StudentCard } from './lib/Molecules/Cards/StudentCard.svelte';
 export { default as TalkCard } from './lib/Molecules/Cards/TalkCard.svelte';
+export { default as TopicNavigatorCard } from './lib/Molecules/Cards/TopicNavigatorCard.svelte';
 export { default as VideoCard } from './lib/Molecules/Cards/VideoCard.svelte';
 
 export { default as CardDeck } from './lib/Organisms/CardDeck/CardDeck.svelte';
+export { default as TopicNavigator } from './lib/Organisms/TopicNavigator/TopicNavigator.svelte';
 export { default as UnitCard } from './lib/Organisms/UnitCard/UnitCard.svelte';
