+++ sites/tutors-course-reader/src/pages/Note.svelte
@@ -1,6 +1,6 @@
 <script lang="ts">
   import { getContext, onMount } from "svelte";
-  import TopicNavigatorCard from "../components/cards/TopicNavigatorCard.svelte";
+  import { TopicNavigatorCard } from "tutors-ui";
   import type { AnalyticsService } from "../reader-lib/services/analytics-service";
   import { revealSidebar } from "../stores";
   import * as animateScroll from "svelte-scrollto";
